+++ frontend/src/pages/translations/components/outbound_translations.js
@@ -15,8 +15,7 @@ export default function OutboundTranslations({
     handleCopy,
     handleDelete
 }) {
-    console.log('OutboundTranslations - displayedRules:', displayedRules); // DEBUG
-    
+    const userGroups = JSON.parse(sessionStorage.getItem('userGroups') || '[]');
     return (
         <table style={{ width: '100%', borderCollapse: 'collapse' }}>
             <thead>
@@ -177,6 +176,7 @@ export default function OutboundTranslations({
                         </td>
                         <td style={{ padding: 4, border: '1px solid #ccc' }}>{rule.trns_output_value}</td>
                         <td style={{ padding: 4, border: '1px solid #ccc', textAlign: 'center' }}>
+                            {userGroups.includes("EWATier1") && (
                             <div style={{ display: 'flex', gap: '8px', justifyContent: 'center' }}>
                                 <button
                                     onClick={() => handleEdit(rule)}
@@ -224,6 +224,7 @@ export default function OutboundTranslations({
                                     <FiTrash2 size={18} color="#000000ff" />
                                 </button>
                             </div>
+                            )}
                         </td>
                     </tr>
                 ))}
